+++ vitest.config.ts
@@ -4,5 +4,6 @@ export default defineConfig({
   test: {
     globals: false,
     environment: "node",
+    include: ["src/**/*.test.ts"],
   },
 });